sub_challenge/requirements.txt
docopt==0.6.2
webvtt-py==0.4.6
XlsxWriter==3.0.3
